Moved shouldReturnForbiddenWhenEventUserRoleIsExternal test to its own class to ensure that ordering of the tests run is not the issue.
